back/requirements.txt
click==8.1.3
Flask==2.1.0
Flask-Cors==3.0.10
importlib-metadata==5.1.0
itsdangerous==2.1.2
psycopg2-binary==2.9.5
python-dotenv==0.21.0
six==1.16.0
Werkzeug==2.2.2
zipp==3.10
Jinja2
Faker
MarkupSafe